Remove dead code and rename the files that hid it
Four public types had no reference anywhere in the repository beyond
their own declaration: XLDatabaseMetadata and its only conformer
XLDatabaseMetadataObject, XLTableName, and XLUnionDependency.

GRDBRowAdapter was a duplicate of GRDBValuesAdapter differing only in its
initializer, and only tests constructed it. Merged: GRDBValuesAdapter
gains init(row:). Neither forwarded XLStaticColumnReader's dialectValue,
which the audit flagged as a possible latent bug; it is not. Both are
used only where a custom function reads intrinsic values by position, and
production static-layout decoding reaches XLSQLiteValueReader -- which
does conform -- directly through GRDBRowDecoder. The type now says so.

Validator dead code: five computed capability sets nothing read, and
hasFunction's argumentCount parameter, which no caller ever passed and
which asks the wrong question anyway, since SQLite reports -1 for a
variadic function. delegatedChecks was rebuilt and re-sorted per report
despite being the same list every time; it is a constant now, which
matters because the report's bytes are a determinism gate. The manifest
placeholder analysis carried a largestPhysicalIndex nothing read.

The six `/// TODO: Remove` comments on XLResult's anonymous-result
requirements said the surface was pending removal. It is not: every one
is live through SQLFunctionalSyntax and generated by both macros.
Retiring it is issue #90 and a v2 change. Each now documents what it
does, with the removal tracked where it belongs.

Renames, all navigation traps rather than style:

- SQLInvocationBindings.swift becomes ContextualBindingReference.swift.
  SwiftQLCore has a file of the same name holding a disjoint set of
  types, so opening the wrong one shows none of the symbols you came for.
- Its legacy half -- the reflective type-to-storage mapping the pre-codec
  path depends on -- moves to LegacyValueMetadata.swift with a v2-removal
  banner, so what goes away is visible as a file.
- SQLStaticQueryDescriptor.swift becomes SQLStaticQueryDescriptor+Encoding
  .swift. The descriptor types are in SwiftQLCore; this file holds only
  the two extensions that build one from a rendering.
- The four root-level codec files join Codecs/.
