+++ Sources/SwiftQL/Codecs/LegacyLiteralValueCodec.swift
