+++ Sources/SwiftQL/Codecs/SQLiteNumericDateCodecs.swift
