+++ Sources/SwiftQL/Codecs/UUIDValueCodec.swift
