+++ Sources/SwiftQL/ContextualBindingReference.swift
@@ -1,3 +1,15 @@
+//
+//  ContextualBindingReference.swift
+//  SwiftQL
+//
+//  The contextual binding reference and the request-facade errors around it.
+//
+//  Renamed from `SQLInvocationBindings.swift` (issue #555): SwiftQLCore has a
+//  file of that name holding an entirely disjoint set of types, and two files
+//  one name apart with nothing in common is a navigation trap -- open the wrong
+//  one and every symbol you were looking for is missing.
+//
+
 import Foundation
 
 
@@ -235,109 +247,6 @@ extension GRDBDatabase {
                 actual: actual,
                 context: context
             )
-        }
-    }
-}
-
-
-func _xlLegacyParameterDeclaration<Value>(
-    for type: Value.Type,
-    key: XLBindingKey
-) -> XLParameterDeclaration {
-    let metadata = legacyValueMetadata(for: type)
-    return XLParameterDeclaration(
-        key: key,
-        valueTypeIdentifier: metadata.identifier,
-        valueTypeName: metadata.typeName,
-        nullability: metadata.isOptional ? .nullable : .required,
-        codecIdentity: nil,
-        codingContext: XLValueCodingContext(
-            site: .parameter,
-            path: XLValueCodingPath(key.contextPathComponent)
-        )
-    )
-}
-
-
-protocol _XLOptionalLiteralType {
-    static var wrappedType: Any.Type { get }
-}
-
-
-extension Optional: _XLOptionalLiteralType {
-    static var wrappedType: Any.Type {
-        Wrapped.self
-    }
-}
-
-
-func sqliteStorageClass(
-    for type: Any.Type
-) -> XLSQLiteStorageClass? {
-    if let optional = type as? any _XLOptionalLiteralType.Type {
-        return sqliteStorageClass(for: optional.wrappedType)
-    }
-    if type == Bool.self || type == Int.self {
-        return .integer
-    }
-    if type == Double.self {
-        return .real
-    }
-    if type == String.self {
-        return .text
-    }
-    if type == Data.self {
-        return .blob
-    }
-    return nil
-}
-
-
-func legacyValueMetadata(
-    for type: Any.Type
-) -> (identifier: XLValueTypeIdentifier, typeName: String, isOptional: Bool) {
-    if let optional = type as? any _XLOptionalLiteralType.Type {
-        let wrapped = legacyValueMetadata(for: optional.wrappedType)
-        return (wrapped.identifier, wrapped.typeName, true)
-    }
-
-    let identifier: String
-    if type == Bool.self {
-        identifier = "swift.bool"
-    }
-    else if type == Int.self {
-        identifier = "swift.int"
-    }
-    else if type == Double.self {
-        identifier = "swift.double"
-    }
-    else if type == String.self {
-        identifier = "swift.string"
-    }
-    else if type == Data.self {
-        identifier = "foundation.data"
-    }
-    else {
-        // v1 custom literals have no durable identifier contract. Keep one
-        // explicit compatibility sentinel and retain the reflected spelling
-        // only in the diagnostic `valueTypeName` field.
-        identifier = "swiftql.v1.legacy-custom"
-    }
-    return (
-        XLValueTypeIdentifier(rawValue: identifier),
-        String(reflecting: type),
-        false
-    )
-}
-
-
-private extension XLBindingKey {
-    var contextPathComponent: String {
-        switch self {
-        case .named(let name):
-            return name
-        case .indexed(let index):
-            return String(index)
         }
     }
 }
